- refactor PointCloudsAndLaserScans to two separate topic extensions - pull out shared functions to separate file `pointScanUtils.ts`
