+++ packages/studio-base/src/panels/ThreeDeeRender/Renderer.ts
@@ -50,10 +50,11 @@ import { CoreSettings } from "./renderables/CoreSettings";
 import { FrameAxes, LayerSettingsTransform } from "./renderables/FrameAxes";
 import { Grids } from "./renderables/Grids";
 import { Images } from "./renderables/Images";
+import { LaserScans } from "./renderables/LaserScans";
 import { Markers } from "./renderables/Markers";
 import { MeasurementTool } from "./renderables/MeasurementTool";
 import { OccupancyGrids } from "./renderables/OccupancyGrids";
-import { PointCloudsAndLaserScans } from "./renderables/PointCloudsAndLaserScans";
+import { PointClouds } from "./renderables/PointClouds";
 import { Polygons } from "./renderables/Polygons";
 import { PoseArrays } from "./renderables/PoseArrays";
 import { Poses } from "./renderables/Poses";
@@ -474,13 +475,14 @@ export class Renderer extends EventEmitter<RendererEvents> {
     this.addSceneExtension(new Markers(this));
     this.addSceneExtension(new FoxgloveSceneEntities(this));
     this.addSceneExtension(new FoxgloveGrid(this));
+    this.addSceneExtension(new LaserScans(this));
     this.addSceneExtension(new OccupancyGrids(this));
-    this.addSceneExtension(new PointCloudsAndLaserScans(this));
-    this.addSceneExtension(new VelodyneScans(this));
+    this.addSceneExtension(new PointClouds(this));
     this.addSceneExtension(new Polygons(this));
     this.addSceneExtension(new Poses(this));
     this.addSceneExtension(new PoseArrays(this));
     this.addSceneExtension(new Urdfs(this));
+    this.addSceneExtension(new VelodyneScans(this));
     this.addSceneExtension(this.measurementTool);
     this.addSceneExtension(this.publishClickTool);
