+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/LaserScans.ts
@@ -1,0 +1,657 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import * as THREE from "three";
+
+import { Time, toNanoSec } from "@foxglove/rostime";
+import { LaserScan as FoxgloveLaserScan } from "@foxglove/schemas";
+import { SettingsTreeAction } from "@foxglove/studio";
+import type { RosObject, RosValue } from "@foxglove/studio-base/players/types";
+import { emptyPose } from "@foxglove/studio-base/util/Pose";
+
+import { DynamicBufferGeometry } from "../DynamicBufferGeometry";
+import { BaseUserData, Renderable } from "../Renderable";
+import { Renderer } from "../Renderer";
+import { PartialMessage, PartialMessageEvent, SceneExtension } from "../SceneExtension";
+import { SettingsTreeEntry, SettingsTreeNodeWithActionHandler } from "../SettingsManager";
+import { LASERSCAN_DATATYPES as FOXGLOVE_LASERSCAN_DATATYPES } from "../foxglove";
+import { normalizeFloat32Array, normalizeTime, normalizePose } from "../normalizeMessages";
+import { LASERSCAN_DATATYPES as ROS_LASERSCAN_DATATYPES, LaserScan as RosLaserScan } from "../ros";
+import { topicIsConvertibleToSchema } from "../topicIsConvertibleToSchema";
+import { MAX_DURATION, Pose } from "../transforms";
+import { updatePose } from "../updatePose";
+import { colorHasTransparency, getColorConverter } from "./pointClouds/colors";
+import {
+  createGeometry,
+  createPoints,
+  DEFAULT_POINT_SCAN_SETTINGS,
+  LayerSettingsPointScan,
+  pointScansSettingsNode,
+  PointsAtTime,
+} from "./pointScanUtils";
+import { missingTransformMessage, MISSING_TRANSFORM } from "./transforms";
+
+type LayerSettingsLaserScan = LayerSettingsPointScan;
+const DEFAULT_SETTINGS = DEFAULT_POINT_SCAN_SETTINGS;
+
+type NormalizedLaserScan = {
+  timestamp: Time;
+  frame_id: string;
+  pose: Pose;
+  start_angle: number;
+  end_angle: number;
+  range_min: number;
+  range_max: number;
+  ranges: Float32Array;
+  intensities: Float32Array;
+};
+
+type LaserScanUserData = BaseUserData & {
+  settings: LayerSettingsLaserScan;
+  topic: string;
+  laserScan?: NormalizedLaserScan;
+  originalMessage: Record<string, RosValue> | undefined;
+  pointsHistory: PointsAtTime[];
+  material: LaserScanMaterial;
+  pickingMaterial: LaserScanMaterial;
+  instancePickingMaterial: LaserScanInstancePickingMaterial;
+};
+
+const ALL_LASERSCAN_DATATYPES = new Set<string>([
+  ...FOXGLOVE_LASERSCAN_DATATYPES,
+  ...ROS_LASERSCAN_DATATYPES,
+]);
+
+const INVALID_LASERSCAN = "INVALID_LASERSCAN";
+
+const VEC3_ZERO = new THREE.Vector3();
+
+const tempColor = { r: 0, g: 0, b: 0, a: 0 };
+
+export class LaserScanRenderable extends Renderable<LaserScanUserData> {
+  public override pickableInstances = true;
+
+  public override dispose(): void {
+    this.userData.laserScan = undefined;
+    this.userData.originalMessage = undefined;
+    for (const entry of this.userData.pointsHistory) {
+      entry.points.geometry.dispose();
+    }
+    this.userData.pointsHistory.length = 0;
+    this.userData.material.dispose();
+    this.userData.pickingMaterial.dispose();
+    this.userData.instancePickingMaterial.dispose();
+    super.dispose();
+  }
+
+  public override details(): Record<string, RosValue> {
+    return this.userData.originalMessage ?? {};
+  }
+
+  public override instanceDetails(instanceId: number): Record<string, RosValue> | undefined {
+    if (!this.userData.laserScan) {
+      return undefined;
+    }
+    const range =
+      instanceId >= 0 && instanceId < this.userData.laserScan.ranges.length
+        ? this.userData.laserScan.ranges[instanceId]
+        : undefined;
+    const intensity =
+      instanceId >= 0 && instanceId < this.userData.laserScan.intensities.length
+        ? this.userData.laserScan.intensities[instanceId]
+        : undefined;
+    return { range, intensity };
+  }
+
+  public updateLaserScan(
+    laserScan: NormalizedLaserScan,
+    originalMessage: RosObject | undefined,
+    settings: LayerSettingsLaserScan,
+    receiveTime: bigint,
+  ): void {
+    const messageTime = toNanoSec(laserScan.timestamp);
+    this.userData.receiveTime = receiveTime;
+    this.userData.messageTime = messageTime;
+    this.userData.frameId = this.renderer.normalizeFrameId(laserScan.frame_id);
+    this.userData.laserScan = laserScan;
+    this.userData.originalMessage = originalMessage;
+
+    this.userData.settings = settings;
+    const { colorField } = settings;
+    const { intensities, ranges } = laserScan;
+
+    // Invalid laser scan checks
+    if (intensities.length !== 0 && intensities.length !== ranges.length) {
+      const message = `LaserScan intensities length (${intensities.length}) does not match ranges length (${ranges.length})`;
+      invalidLaserScanError(this.renderer, this, message);
+      return;
+    }
+    if (colorField !== "intensity" && colorField !== "range") {
+      const message = `LaserScan color field must be either 'intensity' or 'range', found '${colorField}'`;
+      invalidLaserScanError(this.renderer, this, message);
+      return;
+    }
+
+    const laserScanMaterial = this.userData.material;
+    const pickingMaterial = this.userData.pickingMaterial;
+
+    const topic = this.userData.topic;
+    const pointsHistory = this.userData.pointsHistory;
+    const isDecay = settings.decayTime > 0;
+    if (isDecay) {
+      // Push a new (empty) entry to the history of points
+      const geometry = createGeometry(topic, THREE.StaticDrawUsage);
+      const points = createPoints(
+        topic,
+        laserScan.pose,
+        geometry,
+        laserScanMaterial,
+        pickingMaterial,
+        undefined,
+      );
+      pointsHistory.push({ receiveTime, messageTime, points });
+      this.add(points);
+    }
+
+    const latestEntry = pointsHistory[pointsHistory.length - 1];
+    if (!latestEntry) {
+      throw new Error(`pointsHistory is empty for ${topic}`);
+    }
+
+    latestEntry.receiveTime = receiveTime;
+    latestEntry.messageTime = messageTime;
+
+    const geometry = latestEntry.points.geometry;
+    geometry.resize(ranges.length);
+    const rangeAttribute = geometry.attributes.position!;
+    const colorAttribute = geometry.attributes.color!;
+    rangeAttribute.set(ranges);
+
+    // Update material uniforms
+    laserScanMaterial.update(settings, laserScan);
+    pickingMaterial.update(settings, laserScan);
+
+    // Determine min/max color values (if needed) and max range
+    let minColorValue = settings.minValue ?? Number.POSITIVE_INFINITY;
+    let maxColorValue = settings.maxValue ?? Number.NEGATIVE_INFINITY;
+    if (settings.minValue == undefined || settings.maxValue == undefined) {
+      let maxRange = 0;
+
+      for (let i = 0; i < ranges.length; i++) {
+        const range = ranges[i]!;
+        if (Number.isFinite(range)) {
+          maxRange = Math.max(maxRange, range);
+        }
+
+        const colorValue = colorField === "range" ? range : intensities[i];
+        if (Number.isFinite(colorValue)) {
+          minColorValue = Math.min(minColorValue, colorValue!);
+          maxColorValue = Math.max(maxColorValue, colorValue!);
+        }
+      }
+      minColorValue = settings.minValue ?? minColorValue;
+      maxColorValue = settings.maxValue ?? maxColorValue;
+
+      // Update the LaserScan bounding sphere
+      latestEntry.points.geometry.boundingSphere ??= new THREE.Sphere();
+      latestEntry.points.geometry.boundingSphere.set(VEC3_ZERO, maxRange);
+      latestEntry.points.frustumCulled = true;
+    } else {
+      latestEntry.points.geometry.boundingSphere = ReactNull;
+      latestEntry.points.frustumCulled = false;
+    }
+
+    // Build a method to convert raw color field values to RGBA
+    const colorConverter =
+      settings.colorMode === "rgba-fields"
+        ? () => NaN // should never happen: rgba-fields is not supported for LaserScans
+        : getColorConverter(
+            settings as typeof settings & { colorMode: typeof settings.colorMode },
+            minColorValue,
+            maxColorValue,
+          );
+
+    // Iterate the point cloud data to update color attribute
+    for (let i = 0; i < ranges.length; i++) {
+      const colorValue = colorField === "range" ? ranges[i]! : intensities[i] ?? 0;
+      colorConverter(tempColor, colorValue);
+      colorAttribute.setXYZW(
+        i,
+        (tempColor.r * 255) | 0,
+        (tempColor.g * 255) | 0,
+        (tempColor.b * 255) | 0,
+        (tempColor.a * 255) | 0,
+      );
+    }
+
+    rangeAttribute.needsUpdate = true;
+    colorAttribute.needsUpdate = true;
+  }
+
+  public startFrame(currentTime: bigint, renderFrameId: string, fixedFrameId: string): void {
+    const path = this.userData.settingsPath;
+
+    this.visible = this.userData.settings.visible;
+    if (!this.visible) {
+      this.renderer.settings.errors.clearPath(path);
+      const pointsHistory = this.userData.pointsHistory;
+      for (const entry of pointsHistory.splice(0, pointsHistory.length - 1)) {
+        entry.points.geometry.dispose();
+        this.remove(entry.points);
+      }
+      return;
+    }
+
+    // Remove expired entries from the history of points when decayTime is enabled
+    const pointsHistory = this.userData.pointsHistory;
+    const decayTime = this.userData.settings.decayTime;
+    const expireTime =
+      decayTime > 0 ? currentTime - BigInt(Math.round(decayTime * 1e9)) : MAX_DURATION;
+    while (pointsHistory.length > 1 && pointsHistory[0]!.receiveTime < expireTime) {
+      const entry = this.userData.pointsHistory.shift()!;
+      this.remove(entry.points);
+      entry.points.geometry.dispose();
+    }
+
+    // Update the pose on each THREE.Points entry
+    let hadTfError = false;
+    for (const entry of pointsHistory) {
+      const srcTime = entry.messageTime;
+      const frameId = this.userData.frameId;
+      const updated = updatePose(
+        entry.points,
+        this.renderer.transformTree,
+        renderFrameId,
+        fixedFrameId,
+        frameId,
+        currentTime,
+        srcTime,
+      );
+      if (!updated && !hadTfError) {
+        const message = missingTransformMessage(renderFrameId, fixedFrameId, frameId);
+        this.renderer.settings.errors.add(path, MISSING_TRANSFORM, message);
+        hadTfError = true;
+      }
+    }
+
+    if (!hadTfError) {
+      this.renderer.settings.errors.remove(path, MISSING_TRANSFORM);
+    }
+
+    // Update the pixeRatio uniform if the current material is a LaserScanMaterial
+    const material = this.userData.material as Partial<LaserScanMaterial>;
+    const pixelRatio = material.uniforms?.pixelRatio;
+    if (pixelRatio) {
+      pixelRatio.value = this.renderer.getPixelRatio();
+    }
+  }
+}
+
+export class LaserScans extends SceneExtension<LaserScanRenderable> {
+  private fieldsByTopic = new Map<string, string[]>();
+
+  public constructor(renderer: Renderer) {
+    super("foxglove.LaserScans", renderer);
+
+    renderer.addSchemaSubscriptions(ROS_LASERSCAN_DATATYPES, this.handleLaserScan);
+    renderer.addSchemaSubscriptions(FOXGLOVE_LASERSCAN_DATATYPES, this.handleLaserScan);
+  }
+
+  public override settingsNodes(): SettingsTreeEntry[] {
+    const configTopics = this.renderer.config.topics;
+    const handler = this.handleSettingsAction;
+    const entries: SettingsTreeEntry[] = [];
+    for (const topic of this.renderer.topics ?? []) {
+      const isLaserScan = topicIsConvertibleToSchema(topic, ALL_LASERSCAN_DATATYPES);
+      if (!isLaserScan) {
+        continue;
+      }
+      const config = (configTopics[topic.name] ?? {}) as Partial<LayerSettingsLaserScan>;
+      const node: SettingsTreeNodeWithActionHandler = pointScansSettingsNode(
+        this.fieldsByTopic,
+        config,
+        topic,
+        "laserscan",
+      );
+      node.handler = handler;
+      entries.push({ path: ["topics", topic.name], node });
+    }
+    return entries;
+  }
+
+  public override startFrame(
+    currentTime: bigint,
+    renderFrameId: string,
+    fixedFrameId: string,
+  ): void {
+    // Do not call super.startFrame() since we handle updatePose() manually.
+    // Instead of updating the pose for each Renderable in this.renderables, we
+    // update the pose of each THREE.Points object in the pointsHistory of each
+    // renderable
+
+    for (const renderable of this.renderables.values()) {
+      renderable.startFrame(currentTime, renderFrameId, fixedFrameId);
+    }
+  }
+
+  public override handleSettingsAction = (action: SettingsTreeAction): void => {
+    const path = action.payload.path;
+    if (action.action !== "update" || path.length !== 3) {
+      return;
+    }
+
+    this.saveSetting(path, action.payload.value);
+
+    // Update the renderable
+    const topicName = path[1]!;
+    const renderable = this.renderables.get(topicName);
+    if (renderable) {
+      const prevSettings = this.renderer.config.topics[topicName] as
+        | Partial<LayerSettingsLaserScan>
+        | undefined;
+      const settings = { ...DEFAULT_SETTINGS, ...prevSettings };
+
+      if (renderable.userData.laserScan) {
+        renderable.updateLaserScan(
+          renderable.userData.laserScan,
+          renderable.userData.originalMessage,
+          settings,
+          renderable.userData.receiveTime,
+        );
+      }
+    }
+  };
+
+  private handleLaserScan = (
+    messageEvent: PartialMessageEvent<RosLaserScan | FoxgloveLaserScan>,
+  ): void => {
+    const topic = messageEvent.topic;
+    const laserScan =
+      "header" in messageEvent.message
+        ? normalizeRosLaserScan(messageEvent.message)
+        : normalizeFoxgloveLaserScan(messageEvent.message as PartialMessage<FoxgloveLaserScan>);
+    const receiveTime = toNanoSec(messageEvent.receiveTime);
+
+    let renderable = this.renderables.get(topic);
+    if (!renderable) {
+      // Set the initial settings from default values merged with any user settings
+      const userSettings = this.renderer.config.topics[topic] as
+        | Partial<LayerSettingsLaserScan>
+        | undefined;
+      const settings = { ...DEFAULT_SETTINGS, ...userSettings };
+      if (settings.colorField == undefined) {
+        settings.colorField = "intensity";
+        settings.colorMode = "colormap";
+        settings.colorMap = "turbo";
+
+        // Update user settings with the newly selected color field
+        this.renderer.updateConfig((draft) => {
+          const updatedUserSettings = { ...userSettings };
+          updatedUserSettings.colorField = settings.colorField;
+          updatedUserSettings.colorMode = settings.colorMode;
+          updatedUserSettings.colorMap = settings.colorMap;
+          draft.topics[topic] = updatedUserSettings;
+        });
+      }
+
+      const geometry = new DynamicBufferGeometry();
+      geometry.name = `${topic}:LaserScan:geometry`;
+      // Three.JS doesn't render anything if there is no attribute named position, so we use the name position for the "range" parameter.
+      geometry.createAttribute("position", Float32Array, 1);
+      geometry.createAttribute("color", Uint8Array, 4, true);
+
+      const material = new LaserScanMaterial();
+      const pickingMaterial = new LaserScanMaterial({ picking: true });
+      const instancePickingMaterial = new LaserScanInstancePickingMaterial();
+      const points = createPoints(
+        topic,
+        laserScan.pose,
+        geometry,
+        material,
+        pickingMaterial,
+        instancePickingMaterial,
+      );
+
+      material.update(settings, laserScan);
+      pickingMaterial.update(settings, laserScan);
+
+      const messageTime = toNanoSec(laserScan.timestamp);
+      renderable = new LaserScanRenderable(topic, this.renderer, {
+        receiveTime,
+        messageTime,
+        frameId: this.renderer.normalizeFrameId(laserScan.frame_id),
+        pose: laserScan.pose,
+        settingsPath: ["topics", topic],
+        settings,
+        topic,
+        laserScan,
+        originalMessage: messageEvent.message as RosObject,
+        pointsHistory: [{ receiveTime, messageTime, points }],
+        material,
+        pickingMaterial,
+        instancePickingMaterial,
+      });
+      renderable.add(points);
+
+      this.add(renderable);
+      this.renderables.set(topic, renderable);
+    }
+
+    renderable.updateLaserScan(
+      laserScan,
+      messageEvent.message as RosObject,
+      renderable.userData.settings,
+      receiveTime,
+    );
+  };
+}
+
+export class LaserScanMaterial extends THREE.RawShaderMaterial {
+  private static MIN_PICKING_POINT_SIZE = 8;
+
+  public constructor({ picking = false }: { picking?: boolean } = {}) {
+    super({
+      vertexShader: `\
+        #version 300 es
+        precision highp float;
+        precision highp int;
+        uniform mat4 projectionMatrix, modelViewMatrix;
+
+        uniform float pointSize;
+        uniform float pixelRatio;
+        uniform float angleMin, angleIncrement;
+        uniform float rangeMin, rangeMax;
+        in float position; // range, but must be named position in order for three.js to render anything
+        in mediump vec4 color;
+        out mediump vec4 vColor;
+        void main() {
+          if (position < rangeMin || position > rangeMax) {
+            gl_PointSize = 0.0;
+            return;
+          }
+          vColor = color;
+          float angle = angleMin + angleIncrement * float(gl_VertexID);
+          vec4 pos = vec4(position * cos(angle), position * sin(angle), 0, 1.0);
+          gl_Position = projectionMatrix * modelViewMatrix * pos;
+          ${
+            picking
+              ? `gl_PointSize = pixelRatio * max(pointSize, ${LaserScanMaterial.MIN_PICKING_POINT_SIZE.toFixed(
+                  1,
+                )});`
+              : "gl_PointSize = pixelRatio * pointSize;"
+          }
+
+        }
+      `,
+      fragmentShader: `\
+        #version 300 es
+        #ifdef GL_FRAGMENT_PRECISION_HIGH
+          precision highp float;
+        #else
+          precision mediump float;
+        #endif
+        uniform bool isCircle;
+        ${picking ? "uniform vec4 objectId;" : "in mediump vec4 vColor;"}
+        out vec4 outColor;
+
+        ${THREE.ShaderChunk.encodings_pars_fragment /* for LinearTosRGB() */}
+
+        void main() {
+          if (isCircle) {
+            vec2 cxy = 2.0 * gl_PointCoord - 1.0;
+            if (dot(cxy, cxy) > 1.0) { discard; }
+          }
+          ${picking ? "outColor = objectId;" : "outColor = LinearTosRGB(vColor);"}
+        }
+      `,
+    });
+    this.uniforms = {
+      isCircle: { value: false },
+      pointSize: { value: 1 },
+      pixelRatio: { value: 1 },
+      angleMin: { value: NaN },
+      angleIncrement: { value: NaN },
+      rangeMin: { value: NaN },
+      rangeMax: { value: NaN },
+    };
+    if (picking) {
+      this.uniforms.objectId = { value: [NaN, NaN, NaN, NaN] };
+    }
+  }
+
+  public update(settings: LayerSettingsLaserScan, laserScan: NormalizedLaserScan): void {
+    this.uniforms.isCircle!.value = settings.pointShape === "circle";
+    this.uniforms.pointSize!.value = settings.pointSize;
+    this.uniforms.angleMin!.value = laserScan.start_angle;
+    this.uniforms.angleIncrement!.value =
+      (laserScan.end_angle - laserScan.start_angle) / (laserScan.ranges.length - 1);
+    this.uniforms.rangeMin!.value = laserScan.range_min;
+    this.uniforms.rangeMax!.value = laserScan.range_max;
+    this.uniformsNeedUpdate = true;
+
+    const transparent = colorHasTransparency(settings);
+    if (transparent !== this.transparent) {
+      this.transparent = transparent;
+      this.depthWrite = !this.transparent;
+      this.needsUpdate = true;
+    }
+  }
+}
+
+class LaserScanInstancePickingMaterial extends THREE.RawShaderMaterial {
+  private static MIN_PICKING_POINT_SIZE = 8;
+
+  public constructor() {
+    const minPointSize = LaserScanInstancePickingMaterial.MIN_PICKING_POINT_SIZE.toFixed(1);
+    super({
+      vertexShader: `\
+        #version 300 es
+        precision highp float;
+        precision highp int;
+        uniform mat4 projectionMatrix, modelViewMatrix;
+
+        uniform float pointSize;
+        uniform float pixelRatio;
+        uniform float angleMin, angleIncrement;
+        uniform float rangeMin, rangeMax;
+        in float position; // range, but must be named position in order for three.js to render anything
+        varying vec4 objectId;
+        void main() {
+          if (position < rangeMin || position > rangeMax) {
+            gl_PointSize = 0.0;
+            return;
+          }
+          objectId = vec4(
+            float((gl_VertexID >> 24) & 255) / 255.0,
+            float((gl_VertexID >> 16) & 255) / 255.0,
+            float((gl_VertexID >> 8) & 255) / 255.0,
+            float(gl_VertexID & 255) / 255.0);
+          float angle = angleMin + angleIncrement * float(gl_VertexID);
+          vec4 pos = vec4(position * cos(angle), position * sin(angle), 0, 1.0);
+          gl_Position = projectionMatrix * modelViewMatrix * pos;
+          gl_PointSize = pixelRatio * max(pointSize, ${minPointSize});
+        }
+      `,
+      fragmentShader: `\
+        #version 300 es
+        #ifdef GL_FRAGMENT_PRECISION_HIGH
+          precision highp float;
+        #else
+          precision mediump float;
+        #endif
+        uniform bool isCircle;
+        varying vec4 objectId;
+        out vec4 outColor;
+
+        void main() {
+          if (isCircle) {
+            vec2 cxy = 2.0 * gl_PointCoord - 1.0;
+            if (dot(cxy, cxy) > 1.0) { discard; }
+          }
+          outColor = objectId;
+        }
+      `,
+    });
+    this.uniforms = {
+      isCircle: { value: false },
+      pointSize: { value: 1 },
+      pixelRatio: { value: 1 },
+      angleMin: { value: NaN },
+      angleIncrement: { value: NaN },
+      rangeMin: { value: NaN },
+      rangeMax: { value: NaN },
+    };
+  }
+
+  public update(settings: LayerSettingsLaserScan, laserScan: RosLaserScan): void {
+    this.uniforms.isCircle!.value = settings.pointShape === "circle";
+    this.uniforms.pointSize!.value = settings.pointSize;
+    this.uniforms.angleMin!.value = laserScan.angle_min;
+    this.uniforms.angleIncrement!.value = laserScan.angle_increment;
+    this.uniforms.rangeMin!.value = laserScan.range_min;
+    this.uniforms.rangeMax!.value = laserScan.range_max;
+    this.uniformsNeedUpdate = true;
+  }
+}
+
+function invalidLaserScanError(
+  renderer: Renderer,
+  renderable: LaserScanRenderable,
+  message: string,
+): void {
+  renderer.settings.errors.addToTopic(renderable.userData.topic, INVALID_LASERSCAN, message);
+  const pointsHistory = renderable.userData.pointsHistory;
+  const lastEntry = pointsHistory[pointsHistory.length - 1];
+  lastEntry?.points.geometry.resize(0);
+}
+
+function normalizeFoxgloveLaserScan(
+  message: PartialMessage<FoxgloveLaserScan>,
+): NormalizedLaserScan {
+  return {
+    timestamp: normalizeTime(message.timestamp),
+    frame_id: message.frame_id ?? "",
+    pose: normalizePose(message.pose),
+    start_angle: message.start_angle ?? 0,
+    end_angle: message.end_angle ?? 0,
+    range_min: -Infinity,
+    range_max: Infinity,
+    ranges: normalizeFloat32Array(message.ranges),
+    intensities: normalizeFloat32Array(message.intensities),
+  };
+}
+
+function normalizeRosLaserScan(message: PartialMessage<RosLaserScan>): NormalizedLaserScan {
+  return {
+    timestamp: normalizeTime(message.header?.stamp),
+    frame_id: message.header?.frame_id ?? "",
+    pose: emptyPose(),
+    start_angle: message.angle_min ?? 0,
+    end_angle: message.angle_max ?? 0,
+    range_min: message.range_min ?? -Infinity,
+    range_max: message.range_max ?? Infinity,
+    ranges: normalizeFloat32Array(message.ranges),
+    intensities: normalizeFloat32Array(message.intensities),
+  };
+}
