+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/pointScanUtils.ts
@@ -1,0 +1,343 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import * as THREE from "three";
+
+import { PackedElementField, PointCloud } from "@foxglove/schemas";
+import { SettingsTreeNode, Topic } from "@foxglove/studio";
+import { DynamicBufferGeometry } from "@foxglove/studio-base/panels/ThreeDeeRender/DynamicBufferGeometry";
+import { rgbaToCssString } from "@foxglove/studio-base/panels/ThreeDeeRender/color";
+import { isSupportedField } from "@foxglove/studio-base/panels/ThreeDeeRender/renderables/pointClouds/fieldReaders";
+import { BaseSettings } from "@foxglove/studio-base/panels/ThreeDeeRender/settings";
+import { Pose } from "@foxglove/studio-base/panels/ThreeDeeRender/transforms";
+
+import { POINTCLOUD_DATATYPES as FOXGLOVE_POINTCLOUD_DATATYPES } from "../foxglove";
+import { PointCloud2, POINTCLOUD_DATATYPES as ROS_POINTCLOUD_DATATYPES, PointField } from "../ros";
+import { LaserScanMaterial } from "./LaserScans";
+import {
+  baseColorModeSettingsNode,
+  colorHasTransparency,
+  ColorModeSettings,
+  FS_SRGB_TO_LINEAR,
+  INTENSITY_FIELDS,
+  RGBA_PACKED_FIELDS,
+} from "./pointClouds/colors";
+
+export type LayerSettingsPointScan = BaseSettings &
+  ColorModeSettings & {
+    pointSize: number;
+    pointShape: "circle" | "square";
+    decayTime: number;
+  };
+
+const DEFAULT_POINT_SIZE = 1.5;
+const DEFAULT_POINT_SHAPE = "circle";
+const DEFAULT_COLOR_MAP = "turbo";
+const DEFAULT_FLAT_COLOR = { r: 1, g: 1, b: 1, a: 1 };
+const DEFAULT_MIN_COLOR = { r: 100 / 255, g: 47 / 255, b: 105 / 255, a: 1 };
+const DEFAULT_MAX_COLOR = { r: 227 / 255, g: 177 / 255, b: 135 / 255, a: 1 };
+
+// used by LaserScans, VelodyneScans, and PointClouds
+export const DEFAULT_POINT_SCAN_SETTINGS: LayerSettingsPointScan = {
+  visible: false,
+  frameLocked: false,
+  pointSize: DEFAULT_POINT_SIZE,
+  pointShape: DEFAULT_POINT_SHAPE,
+  decayTime: 0,
+  colorMode: "flat",
+  flatColor: rgbaToCssString(DEFAULT_FLAT_COLOR),
+  colorField: undefined,
+  gradient: [rgbaToCssString(DEFAULT_MIN_COLOR), rgbaToCssString(DEFAULT_MAX_COLOR)],
+  colorMap: DEFAULT_COLOR_MAP,
+  explicitAlpha: 1,
+  minValue: undefined,
+  maxValue: undefined,
+};
+
+export const LASERSCAN_FIELDS = ["range", "intensity"];
+const POINTCLOUD_REQUIRED_FIELDS = ["x", "y", "z"];
+export const POINT_SHAPE_OPTIONS = [
+  { label: "Circle", value: "circle" },
+  { label: "Square", value: "square" },
+];
+
+/**
+ * Creates settings node for Point cloud and scan topics
+ * @param fieldsByTopic - fields by topic
+ * @param config - current topic settings
+ * @param topic - topic string to get settings node for
+ * @param kind - pointcloud, laserscan, or velodynescan
+ * @param defaultSettings - (optional) default settings to use
+ * @returns  - settings node for the topic
+ */
+export function pointScansSettingsNode(
+  fieldsByTopic: Map<string, string[]>,
+  config: Partial<LayerSettingsPointScan>,
+  topic: Topic,
+  kind: "pointcloud" | "laserscan" | "velodynescan",
+  defaultSettings: LayerSettingsPointScan = DEFAULT_POINT_SCAN_SETTINGS,
+): SettingsTreeNode {
+  const msgFields =
+    kind === "laserscan"
+      ? LASERSCAN_FIELDS
+      : fieldsByTopic.get(topic.name) ?? POINTCLOUD_REQUIRED_FIELDS;
+  const pointSize = config.pointSize;
+  const pointShape = config.pointShape ?? "circle";
+  const decayTime = config.decayTime;
+
+  const node = baseColorModeSettingsNode(msgFields, config, topic, defaultSettings, {
+    supportsPackedRgbModes: ROS_POINTCLOUD_DATATYPES.has(topic.schemaName),
+    supportsRgbaFieldsMode: FOXGLOVE_POINTCLOUD_DATATYPES.has(topic.schemaName),
+  });
+  node.fields = {
+    pointSize: {
+      label: "Point size",
+      input: "number",
+      step: 1,
+      placeholder: "2",
+      precision: 2,
+      value: pointSize,
+    },
+    pointShape: {
+      label: "Point shape",
+      input: "select",
+      options: POINT_SHAPE_OPTIONS,
+      value: pointShape,
+    },
+    decayTime: {
+      label: "Decay time",
+      input: "number",
+      step: 0.5,
+      placeholder: "0 seconds",
+      min: 0,
+      precision: 3,
+      value: decayTime,
+    },
+    ...node.fields,
+  };
+
+  node.icon = kind === "laserscan" ? "Radar" : "Points";
+  return node;
+}
+
+/**
+ * Selects optimal color field for settings given message
+ * @param output - settings object to apply auto selection of colorfield to
+ * @param pointCloud - point cloud message
+ * @param { supportsPackedRgbModes } - whether or not the point cloud message supports packed rgb modes
+ * @returns - changes output object to have desired color field selected
+ */
+export function autoSelectColorField(
+  output: LayerSettingsPointScan,
+  pointCloud: PointCloud | PointCloud2,
+  { supportsPackedRgbModes }: { supportsPackedRgbModes: boolean },
+): void {
+  // Prefer color fields first
+  if (supportsPackedRgbModes) {
+    for (const field of pointCloud.fields) {
+      if (!isSupportedField(field)) {
+        continue;
+      }
+      const fieldNameLower = field.name.toLowerCase();
+      if (RGBA_PACKED_FIELDS.has(fieldNameLower)) {
+        output.colorField = field.name;
+        switch (fieldNameLower) {
+          case "rgb":
+            output.colorMode = "rgb";
+            break;
+          default:
+          case "rgba":
+            output.colorMode = "rgba";
+            break;
+        }
+        return;
+      }
+    }
+  }
+
+  // Intensity fields are second priority
+  for (const field of pointCloud.fields) {
+    if (!isSupportedField(field)) {
+      continue;
+    }
+    if (INTENSITY_FIELDS.has(field.name)) {
+      output.colorField = field.name;
+      output.colorMode = "colormap";
+      output.colorMap = "turbo";
+      return;
+    }
+  }
+
+  // Fall back to using the first point cloud field
+  const firstField = (pointCloud.fields as readonly (PackedElementField | PointField)[]).find(
+    (field) => isSupportedField(field),
+  );
+  if (firstField != undefined) {
+    output.colorField = firstField.name;
+    output.colorMode = "colormap";
+    output.colorMap = "turbo";
+    return;
+  }
+}
+/**
+ * Creates a THREE.Points object for a point cloud and scan messages
+ * @param topic - topic name string for naming geometry
+ * @param usage - THREE draw usage (ex: THREE.StaticDrawUsage)
+ * @returns
+ */
+export function createGeometry(topic: string, usage: THREE.Usage): DynamicBufferGeometry {
+  const geometry = new DynamicBufferGeometry(usage);
+  geometry.name = `${topic}:PointScans:geometry`;
+  geometry.createAttribute("position", Float32Array, 3);
+  geometry.createAttribute("color", Uint8Array, 4, true);
+  return geometry;
+}
+
+type Material = THREE.PointsMaterial | LaserScanMaterial;
+type Points = THREE.Points<DynamicBufferGeometry, Material>;
+export type PointsAtTime = { receiveTime: bigint; messageTime: bigint; points: Points };
+
+export function pointCloudColorEncoding<T extends LayerSettingsPointScan>(
+  settings: T,
+): "srgb" | "linear" {
+  switch (settings.colorMode) {
+    case "flat":
+    case "colormap":
+    case "gradient":
+      // converted to linear by getColorConverter
+      return "linear";
+    case "rgb":
+    case "rgba":
+    case "rgba-fields":
+      return "srgb";
+  }
+}
+
+export function createPoints(
+  topic: string,
+  pose: Pose,
+  geometry: DynamicBufferGeometry,
+  material: Material,
+  pickingMaterial: THREE.Material,
+  instancePickingMaterial: THREE.Material | undefined,
+): Points {
+  const points = new THREE.Points<DynamicBufferGeometry, Material>(geometry, material);
+  // We don't calculate the bounding sphere for points, so frustum culling is disabled
+  points.frustumCulled = false;
+  points.name = `${topic}:PointCloud:points`;
+  points.userData = {
+    pickingMaterial,
+    instancePickingMaterial,
+    pose,
+  };
+  return points;
+}
+
+// Fragment shader chunk to convert sRGB to linear RGB
+const FS_POINTCLOUD_SRGB_TO_LINEAR = /* glsl */ `
+outgoingLight = sRGBToLinear(outgoingLight);
+`;
+
+// Fragment shader chunk to render a GL_POINT as a circle
+const FS_POINTCLOUD_CIRCLE = /* glsl */ `
+vec2 cxy = 2.0 * gl_PointCoord - 1.0;
+if (dot(cxy, cxy) > 1.0) { discard; }
+`;
+
+export function pointCloudMaterial<T extends LayerSettingsPointScan>(
+  settings: T,
+): THREE.PointsMaterial {
+  const transparent = colorHasTransparency(settings);
+  const encoding = pointCloudColorEncoding(settings);
+  const scale = settings.pointSize;
+  const shape = settings.pointShape;
+
+  const material = new THREE.PointsMaterial({
+    vertexColors: true,
+    size: scale,
+    sizeAttenuation: false,
+    transparent,
+    // The sorting issues caused by writing semi-transparent pixels to the depth buffer are less
+    // distracting for point clouds than the self-sorting artifacts when depth writing is disabled
+    depthWrite: true,
+  });
+
+  // Tell three.js to recompile the shader when `shape` or `encoding` change
+  material.customProgramCacheKey = () => `${shape}-${encoding}`;
+  material.onBeforeCompile = (shader) => {
+    const SEARCH = "#include <output_fragment>";
+    if (shape === "circle") {
+      // Patch the fragment shader to render points as circles
+      shader.fragmentShader = shader.fragmentShader.replace(SEARCH, FS_POINTCLOUD_CIRCLE + SEARCH);
+    }
+    if (encoding === "srgb") {
+      // Patch the fragment shader to add sRGB->linear color conversion
+      shader.fragmentShader =
+        FS_SRGB_TO_LINEAR +
+        shader.fragmentShader.replace(SEARCH, FS_POINTCLOUD_SRGB_TO_LINEAR + SEARCH);
+    }
+  };
+
+  return material;
+}
+
+export function createPickingMaterial<T extends LayerSettingsPointScan>(
+  settings: T,
+): THREE.ShaderMaterial {
+  const MIN_PICKING_POINT_SIZE = 8;
+
+  // Use a custom shader for picking that sets a minimum point size to make
+  // individual points easier to click on
+  const pointSize = Math.max(settings.pointSize, MIN_PICKING_POINT_SIZE);
+  return new THREE.ShaderMaterial({
+    vertexShader: /* glsl */ `
+      uniform float pointSize;
+      void main() {
+        gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+        gl_PointSize = pointSize;
+      }
+    `,
+    fragmentShader: /* glsl */ `
+      uniform vec4 objectId;
+      void main() {
+        gl_FragColor = objectId;
+      }
+    `,
+    side: THREE.DoubleSide,
+    uniforms: { pointSize: { value: pointSize }, objectId: { value: [NaN, NaN, NaN, NaN] } },
+  });
+}
+
+export function createInstancePickingMaterial<T extends LayerSettingsPointScan>(
+  settings: T,
+): THREE.ShaderMaterial {
+  const MIN_PICKING_POINT_SIZE = 8;
+
+  // Use a custom shader for picking that sets a minimum point size to make
+  // individual points easier to click on
+  const pointSize = Math.max(settings.pointSize, MIN_PICKING_POINT_SIZE);
+  return new THREE.ShaderMaterial({
+    vertexShader: /* glsl */ `
+        uniform float pointSize;
+        varying vec4 objectId;
+        void main() {
+          objectId = vec4(
+            float((gl_VertexID >> 24) & 255) / 255.0,
+            float((gl_VertexID >> 16) & 255) / 255.0,
+            float((gl_VertexID >> 8) & 255) / 255.0,
+            float(gl_VertexID & 255) / 255.0);
+          gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
+          gl_PointSize = pointSize;
+        }
+      `,
+    fragmentShader: /* glsl */ `
+        varying vec4 objectId;
+        void main() {
+          gl_FragColor = objectId;
+        }
+      `,
+    side: THREE.DoubleSide,
+    uniforms: { pointSize: { value: pointSize } },
+  });
+}
